Adds authorization to slides controller

+++ app/controllers/api/slides_controller.rb
@@ -1,5 +1,9 @@
 class Api::SlidesController < ApiController
   def create
+    if !owns_slider?(slide_params[:slider_id])
+      return render json: { errors: 'Forbidden' }, status: 401
+    end
+
     slide = Slide.new(slide_params)
     if slide.save
       render json: slide
@@ -10,6 +14,11 @@ class Api::SlidesController < ApiController
 
   def update
     slide = Slide.find(params[:id])
+
+    if !owns_slide?(slide)
+      return render json: { errors: 'Forbidden' }, status: 401
+    end
+
     if slide.update(slide_params)
       render json: slide
     else
@@ -19,6 +28,11 @@ class Api::SlidesController < ApiController
 
   def destroy
     slide = Slide.find(params[:id])
+
+    if !owns_slide?(slide)
+      return render json: { errors: 'Forbidden' }, status: 401
+    end
+
     if slide.destroy
       render json: slide
     else
@@ -37,5 +51,14 @@ class Api::SlidesController < ApiController
 
   def slide_params
     params.require(:slide).permit(:content, :weight, :image, :slider_id)
+  end
+
+  def owns_slider?(slider_id)
+    slider = Slider.find(slider_id)
+    @current_user.id == slider.user_id
+  end
+
+  def owns_slide?(slide)
+    @current_user.id == slide.slider.user_id
   end
 end
